fix(lib): Include hash in assets name
When using css inlining and multiple output formats (e.g. esm and cjs) the assets order is not deterministic.
This causes issues if the assets are named the same, e.g. if you have multiple "index" entry points the assets will be called index1, index2...
So when not using hashes the last output format might override the assets with a changed order, causing the previous output formats to be invalid.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>